Folders: _notes, CSS, JS, materiale and html files
Files with need of help: editoria_porcoddiodifrecce : This was a test for creating arrows to switch from the works inside "portfolio", they have to be fixed on the screen while you scroll, but the right arrow is a bit of a problem. Moreover position: fixed is not an available option to use because it does not allow commands for scaling from css (es. width: 1;)
Pls help.